// src/__generated__/incidentNotesRelay_query.graphql.ts
export type incidentNotesRelay_query$data = {
  readonly insightNotes: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly author: {
          readonly " $fragmentSpreads": FragmentRefs<"insightUserFields_user">;
        } | null | undefined;
        readonly authorId: string;
        readonly content: string;
        readonly createdAt: Instant;
        readonly id: string;
      };
    }>;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
    };
  };
  readonly " $fragmentType": "incidentNotesRelay_query";
};
export type incidentNotesRelay_query$key = {
  readonly " $data"?: incidentNotesRelay_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"incidentNotesRelay_query">;
};